requirements.txt
pyotp>=2.9.0
python-dotenv>=1.0.0
pyzbar>=0.1.9
Pillow>=10.0.0
opencv-python-headless>=4.8.0
